Add files via upload
•	First run the ifconfig command to get the list of interfaces and from there get the interface ID you want to sniff on.
•	Now add this interface ID to your script from where you want to sniff packets.
•	Run the file sniffPkt.py in Terminal 1 as root user since this needs root privilege to sniff packets. The program will now start sniffing till it gets a packet
Ping google.com from Terminal 2.
•	Terminal 1 where the program was executed now prints information of packet it received. This is called sniffing.
•	One can first see an echo-request sent from src=10.0.2.5 (local) to dst=142.250.199.174(google) and then an echo-reply is got from src=142.250.199.174(google) and dst=10.0.2.5 (local).

Note:
You can get the IP address of google.com by executing dig command

Execute the program sniffPkt.py without root permission.
Root permission is needed for executing sniff programs due to security.
When a one sniffs packets, he/she gets access to all data flowing in the network which might include sensitive data. 
In a multi-user environment, if an unprivileged user gets access to this then they can eavesdrop on sensitive data being transmitted in the network. 
Added to this, we have also set the network interface to “promiscuous” mode when we setup the VM. This is again a privileged operation. This allows for all packets on the network to be captured, not just those intended for a specific device. 
Thus, to avoid potential security risks, sniffing network traffic is limited to root users.