Add files via upload
•	Run the program sniffingAndThenSpoofingB.py as attacker on seed
•	This will now start sniffing the interfaces given in filters
•	Start Host A (10.9.0.5) and ping 10.9.0.99 (a non-existing host on the LAN).
"Destination Host Unreachable", it indicates that the packet was successfully sent out from local machine, but a router along the path towards the destination (10.9.0.99 in this case) could not find a route to that host. 
•	This can happen due to a few reasons, including but not limited to:
o	IP address trying to be reached does not exist on the network.
o	Network with IP address trying to be reached due to network issues or the network doesn't exist.
o	Firewall or security rules preventing the host from sending/receiving packets.
o	To add, if one of trying to spoof ICMP packets (specifically ICMP echo replies) to make a host seem reachable when it's not, strategy's success depends on the network topology and how the routers in the network handle "Destination Unreachable" messages.
o	In some networks, a "Destination Unreachable" message from a router can override and nullify any ICMP echo replies one is trying to spoof, because the router's message is telling the original sender that the host isn't reachable.