Add files via upload
If promiscuous mode has to be turned off, one need’s to specify it in the sniff() function by setting the parameter promisc to 0 or False. 
Note that turning off promiscuous mode means your network interface card will only process packets that are destined for it (based on the MAC address) and will ignore the rest of the traffic on the network. So, you might miss some packets that you'd otherwise capture in promiscuous mode.