Add files via upload
This Python script is an implementation of the traceroute command using the Scapy library. Traceroute is a network diagnostic tool that displays the route taken by packets across a network. It is used to identify the path taken to reach a desired destination from a source.

The script sends Internet Control Message Protocol (ICMP) echo requests (similar to a 'ping') to a destination IP Address ('142.250.199.174' in this case) with incrementing Time-To-Live (TTL) values.

Each router along the path decreases the TTL value by 1 before forwarding the packet. When the TTL value reaches zero, the router sends an ICMP "time exceeded" message back to the source, thereby revealing its IP address.

This script starts with a TTL of 1 and increases it for each iteration, thus discovering and printing each router ("hop") along the path to the final destination until it receives an ICMP echo reply from the target.

If no intermediary router responds within a 2 second timeout, the script assumes a 'hop' occurred.